campus-underground-backend/moodle_service/requirements.txt
requests>=2.28.1
python-dotenv>=0.21.0
Flask>=2.2.2
Flask-Cors>=3.0.10
pydantic>=1.10.2
PyPDF2>=3.0.0
